Create CountInversions.py
Inversion Count for an array indicates – how far (or close) the array is from being sorted(in increasing order).
If array is already sorted then inversion count is 0. 
If array is sorted in reverse order that inversion count is the maximum. 
Basically, two elements A[i] and A[j] form an inversion if A[i] > A[j] and i < j